Move EvaluatorState to state module and rename to EvaluatorInstance.
